requirements.txt
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
tensorflow>=2.6.0
opencv-python>=4.5.3
Pillow>=8.3.1
streamlit>=1.2.0
plotly>=5.3.1
joblib>=1.0.1
googletrans==3.1.0a0
